Change long on BigInteger for @ID and BigDecimal for @amount. Deleted Nested for tests. Added table clear after everyTest. @MockMvc.perform in tests for @put and @get transferred in eternal methods. Deleted @EmployeeApplicationTests.java. Deleted commented code.

+++ src/main/java/com/example/employee/dto/JsonEmployee.java
@@ -3,9 +3,10 @@ package com.example.employee.dto;
 import com.example.employee.entities.Employee;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 
-public record JsonEmployee(Long id,
+public record JsonEmployee(BigInteger id,
                            @NotNull String firstName,
                            @NotNull String lastName,
                            @NotNull String position,
